+++ WordPress/Classes/ViewRelated/Support/SupportTableViewController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import SwiftUI
 import WordPressAuthenticator
 import WordPressShared
 import WordPressUI
